Revert "Attention(23) CUDA (#26466)"
This reverts commit a3e477e0e9b4bbf0d10ac70556cea6d97d95ca36.
