+++ onnxruntime/contrib_ops/cpu/bert/attention_parameters.h
@@ -34,7 +34,6 @@ struct AttentionParameters {
   float mask_filter_value;
   float scale;
   bool use_tf32;
-  bool is_output_bnsh;  // whether the output format is BNSH
   AttentionMaskType mask_type;
   AttentionQkvFormat qkv_format;
 };
