+++ onnxruntime/contrib_ops/cuda/bert/attention_softmax.cu
@@ -165,7 +165,7 @@ __device__ inline void SoftmaxSmall(const int total_sequence_length,
   // Update end position for causal.
   int end = valid_end;
   if (causal) {
-    const int end_causal = (total_sequence_length - sequence_length) + s + 1;
+    const int end_causal = total_sequence_length - sequence_length + s + 1;
     if (end_causal < end) {
       end = end_causal;
     }
@@ -241,7 +241,7 @@ __global__ void SoftmaxLargeKernel(const int total_sequence_length,
   // Update end position for causal.
   int end = valid_end;
   if (causal) {
-    const int end_causal = (total_sequence_length - sequence_length) + s + 1;
+    int end_causal = total_sequence_length - sequence_length + s + 1;
     if (end_causal < end) {
       end = end_causal;
     }
@@ -333,7 +333,7 @@ __global__ void SoftmaxWithRawMaskLargeKernel(const int total_sequence_length,
                            : float(input[index]);
     float thread_data = input_data * rsqrt_head_size;
     if (causal) {
-      int from_index = (total_sequence_length - sequence_length) + s;  // offset in total sequence length.
+      int from_index = total_sequence_length - sequence_length + s;  // offset in total sequence length.
       if (i > from_index) {
         thread_data = -CUDART_INF_F;
       }
@@ -439,7 +439,7 @@ __device__ inline void SoftmaxWithRawMaskSmall(const int total_sequence_length,
     thread_data = float(input[index]) * rsqrt_head_size;
 
     if (causal) {
-      int from_index = (total_sequence_length - sequence_length) + s;  // offset in total sequence length.
+      int from_index = total_sequence_length - sequence_length + s;  // offset in total sequence length.
       if (threadIdx.x > from_index) {
         thread_data = -CUDART_INF_F;
       }
